requirements.txt
requests
pytest
Pillow
websocket-client
websockets
